Extract output export fallback emit helpers
